# Conflicts:
#	src/main/java/com/skillbox/socialnet/controller/DefaultController.java
#	src/main/java/com/skillbox/socialnet/controller/DialogController.java
#	src/main/java/com/skillbox/socialnet/controller/FeedsController.java
#	src/main/java/com/skillbox/socialnet/controller/FriendsController.java
#	src/main/java/com/skillbox/socialnet/controller/PlatformController.java
#	src/main/java/com/skillbox/socialnet/controller/PostController.java
#	src/main/java/com/skillbox/socialnet/controller/ProfileController.java
#	src/main/java/com/skillbox/socialnet/controller/StorageController.java
#	src/main/java/com/skillbox/socialnet/controller/WebSocketController.java
#	src/main/java/com/skillbox/socialnet/exception/ExceptionsHandler.java
#	src/main/java/com/skillbox/socialnet/repository/PersonRepository.java
#	src/main/java/com/skillbox/socialnet/service/CommentService.java
#	src/main/java/com/skillbox/socialnet/service/DialogService.java
#	src/main/java/com/skillbox/socialnet/service/FriendsService.java
#	src/main/java/com/skillbox/socialnet/service/MessageService.java
#	src/main/java/com/skillbox/socialnet/service/PlatformService.java
#	src/main/java/com/skillbox/socialnet/service/StorageLoggingService.java
#	src/main/java/com/skillbox/socialnet/service/StorageService.java
#	src/main/java/com/skillbox/socialnet/service/WebSocketService.java
#	src/main/java/com/skillbox/socialnet/util/LogMaskingConverter.java
#	src/main/java/com/skillbox/socialnet/util/annotation/LogResultAspect.java
#	src/main/java/com/skillbox/socialnet/util/annotation/LoggableAspect.java
#	src/main/resources/static/app.js
#	src/test/resources/sql/account/after.sql
#	src/test/resources/sql/account/before.sql
#	src/test/resources/sql/dialog/before.sql
#	src/test/resources/sql/friend/add_10_persons.sql
#	src/test/resources/sql/friend/after.sql
#	src/test/resources/sql/friend/before.sql
#	src/test/resources/sql/friend/delete_new_persons.sql
#	src/test/resources/sql/person/addPerson.sql
#	src/test/resources/sql/person/blockP1.sql
#	src/test/resources/sql/person/deletePerson.sql
#	src/test/resources/sql/person/deleteTestPost.sql
#	src/test/resources/sql/person/unblockP1.sql
#	src/test/resources/sql/post/addPost.sql
#	src/test/resources/sql/post/deletePost.sql

+++ src/main/java/com/skillbox/socialnet/controller/ProfileController.java
@@ -1,23 +1,23 @@
 package com.skillbox.socialnet.controller;
 
-import com.skillbox.socialnet.model.RQ.PostChangeRQ;
-import com.skillbox.socialnet.model.RQ.UserSearchRQ;
-import com.skillbox.socialnet.model.RQ.UserChangeRQ;
-import com.skillbox.socialnet.model.RS.GeneralListResponse;
-import com.skillbox.socialnet.model.RS.GeneralResponse;
 import com.skillbox.socialnet.model.dto.MessageOkDTO;
 import com.skillbox.socialnet.model.dto.PostDTO;
 import com.skillbox.socialnet.model.dto.UserDTO;
+import com.skillbox.socialnet.model.rq.PostChangeRQ;
+import com.skillbox.socialnet.model.rq.UserChangeRQ;
+import com.skillbox.socialnet.model.rq.UserSearchRQ;
+import com.skillbox.socialnet.model.rs.GeneralListResponse;
+import com.skillbox.socialnet.model.rs.GeneralResponse;
+import com.skillbox.socialnet.service.AuthService;
 import com.skillbox.socialnet.service.PostService;
 import com.skillbox.socialnet.service.UserService;
 import com.skillbox.socialnet.util.ElementPageable;
 import com.skillbox.socialnet.util.annotation.Loggable;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 @CrossOrigin
@@ -28,20 +28,30 @@ public class ProfileController {
 
     private final UserService userService;
     private final PostService postService;
+    private final AuthService authService;
 
     @GetMapping("/me")
     public ResponseEntity<GeneralResponse<UserDTO>> getUser() {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.getUser()));
     }
 
     @PutMapping("/me")
     public ResponseEntity<GeneralResponse<UserDTO>> editUser(
             @RequestBody @Valid UserChangeRQ userChangeRQ) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.editUser(userChangeRQ)));
     }
 
     @DeleteMapping("/me")
     public ResponseEntity<GeneralResponse<String>> deleteUser() {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.deleteUser()));
     }
 
@@ -61,6 +71,9 @@ public class ProfileController {
             @PathVariable(name = "id") int personId,
             @RequestParam(name = "publish_date", defaultValue = "0") long publishDate,
             @RequestBody @Valid PostChangeRQ postChangeRQ) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(postService.addPostToUserWall(personId, publishDate, postChangeRQ)));
     }
 
@@ -88,11 +101,17 @@ public class ProfileController {
 
     @PutMapping("/block/{id}")
     public ResponseEntity<GeneralResponse<MessageOkDTO>> blockUser(@PathVariable int id) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.blockUser(id)));
     }
 
     @DeleteMapping("/block/{id}")
     public ResponseEntity<GeneralResponse<MessageOkDTO>> unblockUser(@PathVariable int id) {
+        if (authService.getPersonFromSecurityContext() == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return ResponseEntity.ok(new GeneralResponse<>(userService.unblockUser(id)));
     }
